refactor(a1): extract _bulk_event_context to core/broker_helpers (BR-CODE-1)
Heavy-helper pass, slice 1. Moves the 348-line _bulk_event_context (the six-
dimension storefront context builder: rivalry / mlb_series / tournament /
weather / holiday / playoff) out of app.py into core/broker_helpers.py as
bulk_event_context. It depends only on the db client + stdlib (re, datetime) +
the playoff helpers already in core.helpers — no app.py globals — so the move is
clean. app.py imports it aliased to the historical _bulk_event_context name;
the single call site is unchanged.

Adds 4 unit tests (fake db, no live DB) to test_core_helpers.py: empty-ids,
all-None default shape, playoff-from-name + rivalry mapping, and series-tag
performer beating the event name.

app.py 8092 -> 7747 LOC. Touched surfaces green (363 passed); the remaining
local failures are env-only native/dep gaps (supabase.Client, pyo3 crypto) in
modules this diff doesn't touch, which CI provides.
